+++ apps/client/src/entities/resume.ts
@@ -2,6 +2,8 @@ import type { User } from '@/entities/user';
 
 export type Resume = {
   id: string;
+  positionTitle: string;
+  companyName: string;
   postId: string;
   author: Pick<
     User,
