+++ apps/client/src/feature/resume/ui/ResumeList.tsx
@@ -1,18 +1,16 @@
 import { useQuery } from '@tanstack/react-query';
 
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from '@/components/card/card.tsx';
+import { Card } from '@/components/card/card.tsx';
 import { useGuardContext } from '@/shared/context/hooks';
 import { ServiceContext } from '@/shared/context/ServiceContext';
 import { TokenContext } from '@/shared/context/TokenContext';
+import { Button } from '@/components/button';
+import { useRouteNavigation } from '@/shared/route/useRouteNavigation.ts';
+import { formatDate } from '@/util/format.ts';
 
 export const ResumeListView = () => {
   const { resumeListData } = useGetResumeList();
+  const { toResumeDetail } = useRouteNavigation()
 
   if (resumeListData === undefined) {
     return <div>로딩중...</div>;
@@ -29,14 +27,18 @@ export const ResumeListView = () => {
   return (
     <div>
       <div className="max-w-screen-lg">
-        <div className="space-y-4 w-3/5">
+        <div className="space-y-4 w-2/3">
           {resumeList.map((resume) => (
-            <Card key={resume.id} className="shadow-md">
-              <CardHeader>
-                <CardTitle>{resume.content}</CardTitle>
-                <CardDescription>{resume.createdAt}</CardDescription>
-              </CardHeader>
-              <CardContent>{resume.author.name}</CardContent>
+            <Card key={resume.id} className="">
+              <Button
+                onClick={() => {
+                  toResumeDetail({ resumeId: resume.id });
+                }}
+                className="flex justify-between w-full h-full p-5"
+              >
+                <span>{resume.companyName}</span>
+                <span className="text-gray-400">{formatDate(resume.createdAt)}</span>
+              </Button>
             </Card>
           ))}
         </div>
